feat: 알림 메세지 추가 및 D-1 추가
