+++ src/notifications/notification.scheduler.ts
@@ -1,6 +1,6 @@
 import { MoveRepository } from '#move/move.repository.js';
 import { NotificationService } from '#notifications/notification.service.js';
-import { getNextWeek } from '#utils/dateUtils.js';
+import { getNextDate, getNextWeek } from '#utils/dateUtils.js';
 import logger from '#utils/logger.js';
 import { Injectable } from '@nestjs/common';
 import { Cron, CronExpression } from '@nestjs/schedule';
@@ -25,15 +25,20 @@ export class NotificationScheduler {
   async createNotificationForTodayAndNextWeek() {
     try {
       const today = new Date();
+      const nextDay = getNextDate(today);
       const nextWeek = getNextWeek(today);
 
       const todaysMoves = await this.moveRepository.findManyByDate(today);
-      const upcomingMoves = await this.moveRepository.findManyByDate(nextWeek);
+      const nextDaysMoves = await this.moveRepository.findManyByDate(nextDay);
+      const nextWeekMoves = await this.moveRepository.findManyByDate(nextWeek);
 
       for (const move of todaysMoves) {
         await this.createMoveNotification(move, NotificationType.D_DAY);
       }
-      for (const move of upcomingMoves) {
+      for (const move of nextDaysMoves) {
+        await this.createMoveNotification(move, NotificationType.D_1);
+      }
+      for (const move of nextWeekMoves) {
         await this.createMoveNotification(move, NotificationType.D_7);
       }
     } catch (error) {
